Flatten chainaim/ dirs: inline pre-rotation validator into validators.py; relocate plugin, scenario, and tests to package roots; repoint entry-point and imports
