--- packages/nest-core/nest_core/chainaim/__init__.py
@@ -1,18 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-"""ChainAim validators for Nanda Town — re-exports only.
-
-Public names are layout-independent; import from here, not from module
-paths, so the package layout can change without touching call sites.
-
-Example::
-
-    from nest_core.chainaim import validate_identity_prerotation
-"""
-
-from __future__ import annotations
-
-from nest_core.chainaim.identity_prerotation_validator import (
-    validate_identity_prerotation,
-)
-
-__all__ = ["validate_identity_prerotation"]
